third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-breakpoints/debugger-set-breakpoint-regex-expected.txt

Tests Debugger.setBreakpointByUrl with isRegex set to true.


Running: testSetNoneOfURLAndRegex
Either url or urlRegex or scriptHash must be specified.

Running: testSetBothURLAndRegex
Either url or urlRegex or scriptHash must be specified.

Running: testSetByRegex
Set timer for test function.
Script execution paused.
Call stack:
    0) testFunction (debugger-set-breakpoint-regex.js:12)
Script execution resumed.
